Fixed inheritance of character in player and hunter class definitions. Created Play level funtion.
The inherited classes now successfully initialise the player and hunter member variables via the character class initialiser list. Constructor overloading was used to enable different input arguments for player and hunter. The play level function takes the input argument of "level" and uses this to create a certain number of hunters and print them on the board. level 5 = 5monsters. To do: make sure you cant get a system error because the level is greater than the board axis.