Added get random board position functionality.
This functionality will be used to randomly generate hunter starting positions on each level to improve replayablility. TO DO: need to add handling code for when random position cannot be found i.e. when board is full.

I also changed the board_pos struct to use ints not chars as I now realise that chars tend to be assumed as string characters.