Added arrow key input functionality. Fixed random number bug. Added comments to keyboard functions
There are now functions to detect arrow key presses on the keyboard which can be used to move the player when a button is pressed, also when a button is held down. Fixed a bug in one of the random number function overloads in which when no inputs are passed into the random number generator, it only generates random numbers between 0 and board dimensions -2.